win64 vs2019 build for libatsc3 in preperation for dtapi integration

+++ src/atsc3_utils.c
@@ -293,7 +293,7 @@ uint32_t block_Seek(block_t* block, int32_t seek_pos) {
 		_ATSC3_UTILS_WARN("block_Seek: block: %p, invalid seek_pos to: %u, clamping to 0",
 				block->p_buffer, seek_pos);
 		block->i_pos = 0;
-	} else if(seek_pos > block->p_size) {
+	} else if(seek_pos > (int64_t)block->p_size) {
 		_ATSC3_UTILS_WARN("block_Seek: block: %p, invalid seek_pos to: %u, clamping to %u",
 				block->p_buffer, seek_pos, block->p_size);
 		block->i_pos = block->p_size;
